+++ Plugins/Events/Events/DMActionEvents.cpp
@@ -559,7 +559,25 @@ int32_t DMActionEvents::HandleDMMessageHook(CNWSMessage *thisPtr, CNWSPlayer *pP
         case MessageDungeonMasterMinor::DumpLocals:
         {
             event += "DUMP_LOCALS";
-            DefaultSignalEvent();
+            auto type = Utils::PeekMessage<int32_t>(thisPtr, 0);
+            std::string target = Utils::ObjectIDToString(Utils::PeekMessage<Types::ObjectID>(thisPtr, 4) & 0x7FFFFFFF);
+
+            auto PushAndSignalDumpLocalsEvent = [&](const std::string& ev) -> bool {
+                Events::PushEventData("TYPE", std::to_string(type));
+                Events::PushEventData("TARGET", target);
+                return Events::SignalEvent(ev, oidDM);
+            };
+
+            if (PushAndSignalDumpLocalsEvent(event + "_BEFORE"))
+            {
+                retVal = m_HandlePlayerToServerDungeonMasterMessageHook->CallOriginal<int32_t>(thisPtr, pPlayer, nMinor, bGroup);
+            }
+            else
+            {
+                retVal = false;
+            }
+
+            PushAndSignalDumpLocalsEvent(event + "_AFTER");
             break;
         }
         case MessageDungeonMasterMinor::GiveGoodAlignment: